Move constants into separate files, format code
