+++ LightStrip.py
@@ -1,5 +1,6 @@
 from threading import Thread
 from time import sleep
+
 from wpilib import AddressableLED, Color8Bit
