+++ constants/ArmConstants.py
@@ -1,48 +1,9 @@
 from enum import Enum
-import math
 
-import wpilib
 import rev
-from wpimath.kinematics import SwerveDrive4Kinematics
-from wpimath.geometry import Translation2d
+import wpilib
 
-
-class FalconConstants:
-    kMotorFreeSpeedRPM = 6380
-    kUnitsPerRotation = 2048
-
-
-class Constants:
-    # !
-    maxSpeed = 1
-    scale_speed = 1
-    rotationCurvature = 2.0
-    rotationDeadzone = 0.1
-
-    isSimulation = False
-
-    period = 0.02
-
-    controller_deadzone = 0.05
-    pilot_controller_id = 0
-    operator_controller_id = 1
-
-    frame_width = 27
-    frame_length = 26
-
-    light_strip_pwm_port = 1
-
-    class NavXPort(Enum):
-        kUSB = 1
-        kSPI = 2
-
-    navxPort = NavXPort.kUSB
-
-
-class GamePieceType(Enum):
-    kCone = 0
-    kCube = 1
-    kEmpty = 2
+from constants.GameConstants import GamePieceType
 
 
 class ArmConstants:
@@ -219,92 +180,3 @@ class ArmConstants:
         class Manual:
             maxAnglePerSecond = 10  # degrees per second
             omegaScale = 1
-
-
-class SwerveConstants:
-    swerveDashboardName = "Swerve Dashboard"
-
-    # to change the robot orientation (which way is front)
-    # or if wheels are in the wrong orientation when rotating but fine when moving forward, change these
-    kDriveKinematics = SwerveDrive4Kinematics(
-        Translation2d(Constants.frame_width / 2, Constants.frame_length / 2),
-        Translation2d(Constants.frame_width / 2, -Constants.frame_length / 2),
-        Translation2d(-Constants.frame_width / 2, -Constants.frame_length / 2),
-        Translation2d(-Constants.frame_width / 2, Constants.frame_length / 2),
-    )
-
-    fl_drive_id = 2
-    fl_turn_id = 3
-    fl_abs_encoder_offset_rad = 2.6546849
-    fl_chassis_angular_offset = -math.pi / 2
-
-    fr_drive_id = 4
-    fr_turn_id = 5
-    fr_abs_encoder_offset_rad = 5.2596858
-    fr_chassis_angular_offset = 0
-
-    bl_drive_id = 6
-    bl_turn_id = 7
-    bl_abs_encoder_offset_rad = 0.5824341
-    bl_chassis_angular_offset = math.pi
-
-    br_drive_id = 8
-    br_turn_id = 9
-    br_abs_encoder_offset_rad = 0.7112656
-    br_chassis_angular_offset = math.pi / 2
-
-    kPTurning = 1
-    # kPTurning = 0.4
-    kITurning = 0
-    kDTurning = 0
-
-    # TODO calibrate
-    kDriveMaxMetersPerSecond = 5.15
-    kDriveMaxAccelerationMetersPerSecond = 3.0
-    kDriveMaxTurnMetersPerSecond = 5.0
-    kDriveMaxTurnAccelerationMetersPerSecond = 3.0
-
-    kPRobotTurn = 0.001
-    kIRobotTurn = 0
-    kDRobotTurn = 0
-
-    inches_to_meters = 39.37
-
-    kDrivingMotorReduction = 5.08
-
-    kWheelDiameterMeters = 3 / inches_to_meters
-    kWheelCircumferenceMeters = kWheelDiameterMeters * math.pi
-    # kWheelRotationsPerMeter = 1 / kWheelCircumference
-    kEncoderPulsesPerRevolution = (
-        kDrivingMotorReduction * FalconConstants.kUnitsPerRotation
-    )  # manually measured as 10532
-    kEncoderPositionPerMeter = kEncoderPulsesPerRevolution / kWheelCircumferenceMeters
-
-    # print("kEncoderPositionPerMeter", kEncoderPositionPerMeter)
-
-    kDriveWheelFreeSpeedRps = (
-        FalconConstants.kMotorFreeSpeedRPM * kWheelCircumferenceMeters
-    ) / kDrivingMotorReduction
-
-    # print("kDriveWheelFreeSpeedRps", kDriveWheelFreeSpeedRps)
-
-    # kDrivingEncoderPositionFactor = (
-    #     kWheelCircumferenceMeters
-    # ) / kDrivingMotorReduction  # meters per encoder position
-
-    # kDrivingEncoderPositionFactor = (
-    #     kWheelCircumferenceMeters * kDrivingMotorReduction
-    # ) / ((FalconConstants.kMotorFreeSpeedRPM / 60) * kEncoderPulsesPerRevolution)
-
-    # kDrivingEncoderVelocityFactor = (
-    #     kWheelCircumferenceMeters / kDrivingMotorReduction
-    # ) / 60.0  # meters per second
-    # # kDrivingEncoderVelocityFactor = 3.135
-
-    # calibrated_vs = (vs / kEncoderPositionPerMeter) * 10 # multiplied by 10 to get m/100ms to m/s
-    # vs = (calibrated_vs / 10) * kEncoderPositionPerMeter
-
-    # TODO tune this once it is on the ground
-    kMaxMeasuredVelocityEncoderUnits = 18000
-    typical_drive_velocity = 0.8
-    kFDriving = (typical_drive_velocity * 1023) / kMaxMeasuredVelocityEncoderUnits
