+++ constants/GameConstants.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class GamePieceType(Enum):
+    kCone = 0
+    kCube = 1
+    kEmpty = 2
